add jwt token

+++ Services/AccountService.js
@@ -2,12 +2,26 @@ const {AccountRepository} = require('../Repository/AccountRepository')
 const {generateUUID, getToDay, getCodeInUUID, getDateFromDatetime} = require('../utils/Generation')
 const {CodeFunction} = require('../Enum/CodeFunction')
 const {AccountMapping} = require('../Mapping/AccountMapping')
+const {RoleAccount} = require("../Enum/RoleAccount");
+const bcrypt = require('bcryptjs');
 
 class AccountService {
 
     static async getAllAccounts() {
         try {
             return await AccountRepository.getAllAccounts();
+        } catch (e) {
+            console.error(e);
+        }
+    }
+
+    static async checkLogin(username, password) {
+        try {
+            let account = await AccountRepository.getAccountByUsername(username);
+            console.log((await bcrypt.compare(password, account.password)));
+            if(!account || !(await bcrypt.compare(password, account.password))) return null;
+
+            return account;
         } catch (e) {
             console.error(e);
         }
@@ -36,6 +50,8 @@ class AccountService {
 
     static async deleteAccountByID(id) {
         try {
+            const accountCheck = AccountRepository.updateAccountByID(id)
+            if(accountCheck.role === RoleAccount.GLOBAL_ADMIN) return null;
             await AccountRepository.deleteAccount(id);
         } catch (e) {
             console.error(e);
